Corrección Error importante (está unos minutos fuera del tiempo, pero es que me di cuenta de un error)

+++ src/main/java/pd/ecp2/miw/webPattern/dao/entities/Vote.java
@@ -3,14 +3,14 @@ public class Vote {
 
     private int id;
 
-    private int vote;
+    private double vote;
 
     private Theme theme;
 
     public Vote() {
     }
 
-    public Vote(int id, int vote, Theme theme) {
+    public Vote(int id, double vote, Theme theme) {
         this.id = id;
         this.vote = vote;
         this.theme = theme;
@@ -20,7 +20,7 @@ public class Vote {
         return id;
     }
 
-    public int getVote() {
+    public double getVote() {
         return vote;
     }
